resolve dep versions from node module
