+++ packages/cli/src/utils/utils.ts
@@ -1,6 +1,10 @@
+import fs from "node:fs";
+import path from "node:path";
+import { createRequire } from "node:module";
 import color from "chalk";
 import { z } from "zod/v4";
 import { error } from "./errors.js";
+import type { PackageJson } from "type-fest";
 
 export function isUrl(path: string) {
 	const result = z.url().safeParse(path);
@@ -44,6 +48,27 @@ export function parseDependency(dep: string) {
 	if (!name || !version) throw error(`Failed to parse dependency: ${dep}`);
 
 	return { name, version };
+}
+
+export function resolveDependencyPkg(cwd: string, depName: string): PackageJson | undefined {
+	const require = createRequire(path.resolve(cwd, "noop.js"));
+
+	const paths = require.resolve.paths(depName);
+	if (!paths) return;
+
+	let pkgPath: string | undefined;
+	for (const nodeModulesPath of paths) {
+		const check = path.join(nodeModulesPath, depName, "package.json");
+		if (fs.existsSync(check)) {
+			pkgPath = check;
+			break;
+		}
+	}
+
+	if (!pkgPath) return;
+
+	const json = fs.readFileSync(pkgPath, "utf8");
+	return JSON.parse(json);
 }
 
 /** Converts a `Set` into an array if its size is greater than 0. Otherwise, `undefined` is returned. */
